pyproject.toml
[tool.poetry]
name = "oobabot"
version = "0.1.0"
description = "A Discord bot which talks to Large Language Model AIs running on oobabooga's text-generation-webui"
authors = ["Christopher Rude <chris@rudesoftware.net>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = ">=2.2.2"
pysbd = ">=0.3.4"
websockets = ">=11"

[tool.poetry.scripts]
oobabot = 'oobabot.oobabot:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
